WIP:emerge-initramfs: Install libgcc_s.so.1
GCC provides a low-level runtime library, libgcc.a or libgcc_s.so.1 on some
platforms. GCC generates calls to routines in this library automatically,
whenever it needs to perform some operation that is too complicated to emit
inline code for. cryptsetup requires this as a runtime dependency.

Note that this work is incomplete; We have works below left:

 1. We need to make ld.so to load libgcc_s.so.1 automatically.
    This can be done by modifying /etc/ld.so.conf or /etc/ld.so.conf.d.